pyproject.toml
[project]
name = "caps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.3",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.20",
    "langchain-google-genai>=2.0.10",
    "langchain-postgres>=0.0.13",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.4.1",
    "sqlmodel>=0.0.24",
    "torch>=2.6.0",
]
